+++ pkg/policy/rule_test.go
@@ -1759,7 +1759,7 @@ func (ds *PolicyTestSuite) TestIngressL4AllowAll(c *C) {
 
 	filter, ok := l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 1)
@@ -1804,7 +1804,7 @@ func (ds *PolicyTestSuite) TestIngressL4AllowAllNamedPort(c *C) {
 
 	filter, ok := l4IngressPolicy["port-80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 0)
+	c.Assert(filter.Port, Equals, uint16(0))
 	c.Assert(filter.PortName, Equals, "port-80")
 	c.Assert(filter.Ingress, Equals, true)
 
@@ -1876,7 +1876,7 @@ func (ds *PolicyTestSuite) TestEgressL4AllowAll(c *C) {
 
 	filter, ok := l4EgressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, false)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 1)
@@ -1932,7 +1932,7 @@ func (ds *PolicyTestSuite) TestEgressL4AllowWorld(c *C) {
 
 	filter, ok := l4EgressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, false)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 1)
@@ -1987,7 +1987,7 @@ func (ds *PolicyTestSuite) TestEgressL4AllowAllEntity(c *C) {
 
 	filter, ok := l4EgressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, false)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 1)
@@ -2159,7 +2159,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filter, ok := l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 2)
@@ -2186,7 +2186,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filterL7, ok := l4IngressPolicy["7000/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filterL7.Port, Equals, 7000)
+	c.Assert(filterL7.Port, Equals, uint16(7000))
 	c.Assert(filterL7.Ingress, Equals, true)
 
 	c.Assert(len(filterL7.PerSelectorPolicies), Equals, 1)
@@ -2266,7 +2266,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filter, ok = l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 2)
@@ -2277,7 +2277,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filterL7, ok = l4IngressPolicy["7000/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filterL7.Port, Equals, 7000)
+	c.Assert(filterL7.Port, Equals, uint16(7000))
 	c.Assert(filterL7.Ingress, Equals, true)
 
 	c.Assert(len(filterL7.PerSelectorPolicies), Equals, 1)
@@ -2331,7 +2331,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filter, ok = l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(filter.L7Parser, Equals, ParserTypeHTTP)
@@ -2385,7 +2385,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filter, ok = l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(filter.L7Parser, Equals, ParserTypeHTTP)
@@ -2444,7 +2444,7 @@ func (ds *PolicyTestSuite) TestL3L4L7Merge(c *C) {
 
 	filter, ok := l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 2)
@@ -2510,7 +2510,7 @@ func (ds *PolicyTestSuite) TestL3L4L7Merge(c *C) {
 
 	filter, ok = l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(filter.L7Parser, Equals, ParserTypeHTTP)
